+++ spynnaker/pyNN/utilities/ranged/spynnaker_ranged_dict.py
@@ -1,10 +1,10 @@
 from spinn_utilities.ranged.range_dictionary import RangeDictionary
-from .spynakker_ranged_list import SpynakkerRangedList
+from .spynnaker_ranged_list import SpynakkerRangedList
 
 
 class SpynakkerRangeDictionary(RangeDictionary):
 
-    def list_factory(self, size=None, value=None, key=None):
+    def list_factory(self, size, value, key):
         """
         Defines which class or subclass of RangedList to use
